requirements.txt
appnope==0.1.4
asttokens==3.0.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
fonttools==4.56.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
-e git+https://github.com/looshua/UM-EECS-556-Final.git@3f030cebd79884211fc10c6a23fd207ab5e58b11#egg=L3RM
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
-e git+https://github.com/looshua/UM-EECS-556-Final.git@3f030cebd79884211fc10c6a23fd207ab5e58b11#egg=PyLR3M
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyzmq==26.3.0
scipy==1.15.2
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13
